Add missing autoconf check for AR
After removing libtool,	there is no more cross-check for AR. Because
some slightly older versions of autoconf don't have AC_PROG_AR, simply
insert the implementation one-liner.

We already tried bumping the minimum autoconf version once in the past,
but apparently that is an issue for a number of distros.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>